fixed weirdness with hs.execute not working in threading macros
In Lua/Fennel, when a function returns multiple values and those are passed to another function that accepts multiple parameters, the parameters get filled with those values.

So in the threading macro, when you just wrote `tonumber`, it was likely interpreting the second and third return values from `hs.execute` as the optional base parameter to `tonumber`, causing unexpected behavior.

Also, hs.execute returns string with line-breaks, had to trim them down